modified utils_mod and utils_net so that :
-- everything related to net_list and Pajek .net file is in utils_net
-- everything related to RESULTS of radatools including .lol, -info_nodes, -info_dists and -info_global
is in utils_modtest_utils_mod + modif

All without the visbrain modification bugs
