+++ graphpype/utils_mod.py
@@ -9,10 +9,11 @@ import pandas as pd
 import numpy as np
 import scipy.sparse as sp
 
+################################## from lol_file
 
-def get_modularity_value_from_lol_file(modularity_file):
+def get_modularity_value_from_lol_file(lol_file):
 
-    with open(modularity_file, 'r') as f:
+    with open(lol_file, 'r') as f:
 
         for line in f.readlines():
 
@@ -30,6 +31,8 @@ def get_modularity_value_from_lol_file(modularity_file):
 
         return -1.0
 
+###################################### reading info files #################################################################
+############### from info-nodes
 
 def get_max_degree_from_node_info_file(info_nodes_file):
     """
@@ -45,6 +48,63 @@ def get_max_degree_from_node_info_file(info_nodes_file):
 
     return df['Degree'].max(), df[df['Degree'] == df['Degree'].max()].Index.values[0], df[df['Degree'] == df['Degree'].max()].Name.values[0]
 
+
+#ead strength from Network_Properties node results
+def get_strength_values_from_info_nodes_file(info_nodes_file):
+
+    from pandas.io.parsers import read_csv
+
+    info_nodes = read_csv(info_nodes_file, sep="\t")
+
+    # print info_nodes
+
+    return info_nodes['Strength'].values
+
+
+def get_strength_pos_values_from_info_nodes_file(info_nodes_file):
+
+    from pandas.io.parsers import read_csv
+
+    info_nodes = read_csv(info_nodes_file, sep="\t")
+
+    print(info_nodes)
+
+    return info_nodes['Strength_Pos']
+
+
+def get_strength_neg_values_from_info_nodes_file(info_nodes_file):
+
+    from pandas.io.parsers import read_csv
+
+    info_nodes = read_csv(info_nodes_file, sep="\t")
+
+    print(info_nodes)
+
+    return info_nodes['Strength_Neg']
+
+
+def get_degree_pos_values_from_info_nodes_file(info_nodes_file):
+
+    from pandas.io.parsers import read_csv
+
+    info_nodes = read_csv(info_nodes_file, sep="\t")
+
+    print(info_nodes)
+
+    return info_nodes['Degree_Pos']
+
+
+def get_degree_neg_values_from_info_nodes_file(info_nodes_file):
+
+    from pandas.io.parsers import read_csv
+
+    info_nodes = read_csv(info_nodes_file, sep="\t")
+
+    print(info_nodes)
+
+    return info_nodes['Degree_Neg']
+
+############## from info_global
 
 def get_values_from_global_info_file(global_info_file):
 
@@ -413,112 +473,84 @@ def get_path_length_from_info_dists_file(info_dists_file):
         sys.exit()
 
     triu_dist_mat = dist_mat[np.triu_indices(dist_mat.shape[0], k=1)]
-    #np.triu(dist_mat,k = 1)
 
     inv_triu_dist_map = 1.0/triu_dist_mat
 
     print(np.mean(inv_triu_dist_map))
 
-    # if np.isinf(np.mean(triu_dist_mat)):
-
-    # print np.mean(triu_dist_mat)
-
-    # print np.mean(inv_triu_dist_map)
-
     return np.mean(triu_dist_mat), np.max(triu_dist_mat), np.mean(inv_triu_dist_map)
 
-    #global_values = {}
+############################################################## modularity
 
-    # with open(global_info_file,'r') as f:
+##################################### Formatting data for external community detection algorithm (radatools) ##############################
+def read_lol_file(lol_file):
 
-    # for line in f.readlines():
+    with open(lol_file, 'r') as f:
 
-    #split_line = line.strip().split(' ')
+        lines = f.readlines()[4:]
 
-    # print split_line
+        nb_elements = int(lines[0].split(': ')[1])
 
-    # if split_line[0] == 'Total' and split_line[1] == 'degree':
+        community_vect = np.empty((nb_elements), dtype=int)
 
-    # print "Found Total degree value line"
+        for i, line in enumerate(lines[3:]):
 
-    # print line.strip().split('\t')[-1]
+            try:
+                nb_nodes, index_nodes = line.split(': ')
+                print(nb_nodes, index_nodes)
 
-    #global_values['Total_degree'] = line.strip().split('\t')[-1]
+                if int(nb_nodes) > 1:
+                    index_nodes = np.array(
+                        list(map(int, index_nodes.split(' '))), dtype=int) - 1
 
-    # elif split_line[0] == 'Average' and split_line[1] == 'degree':
+                    # print i,index_nodes
+                    community_vect[index_nodes] = i
 
-    # print "Found Average degree value line"
+                else:
+                    community_vect[int(index_nodes) - 1] = i
 
-    # print line.strip().split('\t')[-1]
+            except ValueError:
+                print("Warning, error reading lol file ")
 
-    #global_values['Average_degree'] = line.strip().split('\t')[-1]
+        f.close()
 
-    # elif split_line[0] == 'Minimum' and split_line[1] == 'degree':
+    return community_vect
 
-    # print "Found Minimum degree value line"
+# compute modular matrix from sparse matrix and community vect
+def compute_modular_matrix(sparse_matrix, community_vect):
 
-    # print line.strip().split('\t')[-1]
+    mod_mat = np.empty(sparse_matrix.todense().shape)
 
-    #global_values['Minimum_degree'] = line.strip().split('\t')[-1]
+    mod_mat[:] = np.NAN
 
-    # elif split_line[0] == 'Maximum' and split_line[1] == 'degree':
+    for u, v, w in zip(sparse_matrix.row, sparse_matrix.col, sparse_matrix.data):
 
-    # print "Found Maximum degree value line"
+        if (community_vect[u] == community_vect[v]):
 
-    # print line.strip().split('\t')[-1]
+            mod_mat[u, v] = community_vect[u]
+        else:
 
-    #global_values['Maximum_degree'] = line.strip().split('\t')[-1]
+            mod_mat[u, v] = -1
 
-    # elif split_line[0] == 'Assortativity':
+    return mod_mat
 
-    # print "Found Assortativity value line"
+############################################################## Node roles
 
-    # print line.strip().split('\t')[-1]
-
-    #global_values['Assortativity'] = line.strip().split('\t')[-1]
-
-    # return global_values
-
-# Node roles
-
-
-def return_all_Z_com_degree(community_vect, dense_mat):
+def _return_all_Z_com_degree(community_vect, dense_mat):
 
     degree_vect = np.sum(dense_mat != 0, axis=1)
 
-    print("All degree vect")
-
-    print(degree_vect)
-    print(degree_vect.shape)
-
     community_indexes = np.unique(community_vect)
-
-    # print community_indexes
 
     all_Z_com_degree = np.zeros(shape=(community_vect.shape[0]))
 
     for com_index in community_indexes:
 
-        print(np.where(com_index == community_vect))
-
         com_degree = degree_vect[com_index == community_vect]
-
-        print("Commmunity degree vect")
-        print(com_degree)
-        print(com_degree.shape)
 
         if com_degree.shape[0] > 1:
 
-            std_com_degree = np.std(com_degree)
-
-            print(std_com_degree)
-
-            mean_com_degree = np.mean(com_degree)
-
-            print(mean_com_degree)
-            Z_com_degree = (com_degree - mean_com_degree) / std_com_degree
-
-            print(Z_com_degree)
+            Z_com_degree = (com_degree - np.mean(com_degree)) / np.std(com_degree)
 
             all_Z_com_degree[com_index == community_vect] = Z_com_degree
 
@@ -528,46 +560,27 @@ def return_all_Z_com_degree(community_vect, dense_mat):
     return all_Z_com_degree
 
 
-def return_all_participation_coeff(community_vect, dense_mat):
+def _return_all_participation_coeff(community_vect, dense_mat):
 
     degree_vect = np.array(np.sum(dense_mat != 0, axis=1), dtype='float')
 
-    print(degree_vect)
-    print(degree_vect.shape)
-
     community_indexes = np.unique(community_vect)
-
-    print(community_indexes)
 
     all_participation_coeff = np.ones(
         shape=(community_vect.shape[0]), dtype='float')
 
     for com_index in community_indexes:
 
-        print(np.where(com_index == community_vect))
-
-        nod_index = (com_index == community_vect)
-
-        print(np.sum(nod_index, axis=0))
-
-        com_matrix = dense_mat[:, nod_index]
-
-        degree_com_vect = np.sum(com_matrix, axis=1, dtype='float')
-
-        print(degree_com_vect.shape)
+        degree_com_vect = np.sum(dense_mat[:, com_index == community_vect], axis=1, dtype='float')
 
         rel_com_degree = np.square(degree_com_vect/degree_vect)
 
-        print(rel_com_degree)
-
         all_participation_coeff = all_participation_coeff - rel_com_degree
-
-    print(all_participation_coeff)
 
     return all_participation_coeff
 
 
-def return_amaral_roles(all_Z_com_degree, all_participation_coeff):
+def _return_amaral_roles(all_Z_com_degree, all_participation_coeff):
 
     if (all_Z_com_degree.shape[0] != all_participation_coeff.shape[0]):
         print("Warning, all_Z_com_degree %d should have same length as all_participation_coeff %d " % (
@@ -616,7 +629,6 @@ def return_amaral_roles(all_Z_com_degree, all_participation_coeff):
     # provincial hubs
     prov_hubs = np.logical_and(all_participation_coeff < 0.3, hubs == True)
 
-    print(np.sum(prov_hubs, axis=0))
     nod_roles[prov_hubs, 1] = 5
 
     # hub connectors
@@ -630,15 +642,12 @@ def return_amaral_roles(all_Z_com_degree, all_participation_coeff):
     kin_less_hubs = np.logical_and(
         0.75 <= all_participation_coeff, hubs == True)
 
-    print(np.sum(kin_less_hubs, axis=0))
     nod_roles[kin_less_hubs, 1] = 7
-
-    print(nod_roles)
 
     return nod_roles
 
 
-def return_4roles(all_Z_com_degree, all_participation_coeff):
+def _return_4roles(all_Z_com_degree, all_participation_coeff):
 
     if (all_Z_com_degree.shape[0] != all_participation_coeff.shape[0]):
         print("Warning, all_Z_com_degree %d should have same length as all_participation_coeff %d " % (
@@ -650,9 +659,7 @@ def return_4roles(all_Z_com_degree, all_participation_coeff):
     # hubs are at 2,non-hubs are at 1
     hubs = all_Z_com_degree > 1.0
     non_hubs = all_Z_com_degree <= 1.0
-
-    print(np.sum(hubs, axis=0), np.sum(non_hubs, axis=0))
-
+    
     nod_roles[hubs, 0] = 2
     nod_roles[non_hubs, 0] = 1
 
@@ -664,11 +671,11 @@ def return_4roles(all_Z_com_degree, all_participation_coeff):
     # connector nodes
     connector_nodes = 0.3 <= all_participation_coeff
 
-    print(np.sum(provincial_nodes, axis=0), np.sum(connector_nodes, axis=0))
     nod_roles[connector_nodes, 1] = 2
 
-    # print nod_roles
-
+    print("\nNode roles:")
+    print("*Hubs/non-hubs: ",np.sum(hubs, axis=0), np.sum(non_hubs, axis=0))
+    print("*provincials/connectors : ",np.sum(provincial_nodes, axis=0), np.sum(connector_nodes, axis=0))
     return nod_roles
 
 
@@ -678,34 +685,26 @@ def compute_roles(community_vect, sparse_mat, role_type="Amaral_roles"):
 
     dense_mat = sparse_mat.todense()
 
-    print(dense_mat)
-
     undir_dense_mat = dense_mat + np.transpose(dense_mat)
 
     bin_dense_mat = np.array(undir_dense_mat != 0, dtype=int)
 
-    print(bin_dense_mat)
-
     ##################################### within community Z-degree #########################
 
-    all_Z_com_degree = return_all_Z_com_degree(community_vect, bin_dense_mat)
-
-    # print all_Z_com_degree
+    all_Z_com_degree = _return_all_Z_com_degree(community_vect, bin_dense_mat)
 
     ##################################### participation_coeff ###############################
 
-    all_participation_coeff = return_all_participation_coeff(
+    all_participation_coeff = _return_all_participation_coeff(
         community_vect, bin_dense_mat)
-
-    # print all_participation_coeff
 
     if role_type == "Amaral_roles":
 
-        node_roles = return_amaral_roles(
+        node_roles = _return_amaral_roles(
             all_Z_com_degree, all_participation_coeff)
 
     elif role_type == "4roles":
 
-        node_roles = return_4roles(all_Z_com_degree, all_participation_coeff)
+        node_roles = _return_4roles(all_Z_com_degree, all_participation_coeff)
 
     return node_roles, all_Z_com_degree, all_participation_coeff
