+++ graphpype/utils_net.py
@@ -35,43 +35,7 @@ def return_int_net_list(int_mat, min_int=0):
 
     return net_list
 
-##################################### Formatting data for external community detection algorithm (radatools) ##############################
-
-
-def read_lol_file(lol_file):
-
-    with open(lol_file, 'r') as f:
-
-        lines = f.readlines()[4:]
-
-        nb_elements = int(lines[0].split(': ')[1])
-
-        community_vect = np.empty((nb_elements), dtype=int)
-
-        for i, line in enumerate(lines[3:]):
-
-            try:
-                nb_nodes, index_nodes = line.split(': ')
-                print(nb_nodes, index_nodes)
-
-                if int(nb_nodes) > 1:
-                    index_nodes = np.array(
-                        list(map(int, index_nodes.split(' '))), dtype=int) - 1
-
-                    # print i,index_nodes
-                    community_vect[index_nodes] = i
-
-                else:
-                    community_vect[int(index_nodes) - 1] = i
-
-            except ValueError:
-                print("Warning, error reading lol file ")
-
-        f.close()
-
-    return community_vect
-
-
+################################### reading file
 def read_Pajek_corres_nodes(Pajek_net_file):
 
     with open(Pajek_net_file, 'r') as f:
@@ -154,85 +118,7 @@ def read_Pajek_corres_nodes_and_sparse_matrix(Pajek_net_file):
 
     return node_corres, sparse_matrix
 
-# compute modular matrix from sparse matrix and community vect
-
-
-def compute_modular_matrix(sparse_matrix, community_vect):
-
-    mod_mat = np.empty(sparse_matrix.todense().shape)
-
-    mod_mat[:] = np.NAN
-
-    for u, v, w in zip(sparse_matrix.row, sparse_matrix.col, sparse_matrix.data):
-
-        if (community_vect[u] == community_vect[v]):
-
-            mod_mat[u, v] = community_vect[u]
-        else:
-
-            mod_mat[u, v] = -1
-
-    return mod_mat
-
-# read strength from Network_Properties node results
-
-
-def get_strength_values_from_info_nodes_file(info_nodes_file):
-
-    from pandas.io.parsers import read_csv
-
-    info_nodes = read_csv(info_nodes_file, sep="\t")
-
-    # print info_nodes
-
-    return info_nodes['Strength'].values
-
-
-def get_strength_pos_values_from_info_nodes_file(info_nodes_file):
-
-    from pandas.io.parsers import read_csv
-
-    info_nodes = read_csv(info_nodes_file, sep="\t")
-
-    print(info_nodes)
-
-    return info_nodes['Strength_Pos']
-
-
-def get_strength_neg_values_from_info_nodes_file(info_nodes_file):
-
-    from pandas.io.parsers import read_csv
-
-    info_nodes = read_csv(info_nodes_file, sep="\t")
-
-    print(info_nodes)
-
-    return info_nodes['Strength_Neg']
-
-
-def get_degree_pos_values_from_info_nodes_file(info_nodes_file):
-
-    from pandas.io.parsers import read_csv
-
-    info_nodes = read_csv(info_nodes_file, sep="\t")
-
-    print(info_nodes)
-
-    return info_nodes['Degree_Pos']
-
-
-def get_degree_neg_values_from_info_nodes_file(info_nodes_file):
-
-    from pandas.io.parsers import read_csv
-
-    info_nodes = read_csv(info_nodes_file, sep="\t")
-
-    print(info_nodes)
-
-    return info_nodes['Degree_Neg']
-
 ##################################### Formatting data for external community detection algorithm (Louvain_Traag) ##############################
-
 
 def export_Louvain_net_from_list(Z_Louvain_file, Z_list, coords):
